move PredictWorker to workers.monitoring module
